Split the project into smaller source files
