+++ src/cleaners/asMaybe.js
@@ -1,0 +1,9 @@
+export function asMaybe(cleaner, fallback) {
+  return function asMaybe(raw) {
+    try {
+      return cleaner(raw)
+    } catch (e) {
+      return fallback
+    }
+  }
+}
