+++ src/cleaners/asOptional.js
@@ -1,0 +1,5 @@
+export function asOptional(cleaner, fallback) {
+  return function asOptional(raw) {
+    return raw != null ? cleaner(raw) : fallback
+  }
+}
